+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Repository/LikeRepository.java
@@ -9,12 +9,12 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface LikeRepository extends JpaRepository<Like, Integer> {
+public interface LikeRepository extends JpaRepository<Like, Long> {
     boolean existsByUserAndPost(User user, Post post);
     Like findByUserAndPost(User user, Post post);
     @Query("SELECT COUNT(l) FROM Like l WHERE l.post.id = :postId")
-    long countByPostId(int postId);
+    long countByPostId(Long postId);
     @Query("SELECT l.user FROM Like l WHERE l.post.id = :postId")
-    List<User> findUsersByPostId(@Param("postId") int postId);
-    void deleteByPostId(int postId);
+    List<User> findUsersByPostId(@Param("postId") Long postId);
+    void deleteByPostId(Long postId);
 }
