+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Service/PostService.java
@@ -9,15 +9,15 @@ import java.io.IOException;
 import java.util.List;
 
 public interface PostService {
-    int getPostsNrOfUser(int userId);
+    int getPostsNrOfUser(Long userId);
     Post createPost(String jwt, MultipartFile file, String text) throws IOException;
-    void deletePost(int postId, String jwt) throws Exception;
-    List<Post> getAllPostsByUser(int userId);
-    long getLikesCountForPost(int postId);
-    List<UserResponse> getUsersWhoLikedPost(int postId);
-    byte[] getPostMedia(int postId) throws Exception;
-    String getMediaKey(int postId) throws Exception;
+    void deletePost(Long postId, String jwt) throws Exception;
+    List<Post> getAllPostsByUser(Long userId);
+    long getLikesCountForPost(Long postId);
+    List<UserResponse> getUsersWhoLikedPost(Long postId);
+    byte[] getPostMedia(Long postId) throws Exception;
+    String getMediaKey(Long postId) throws Exception;
     List<Post> getPostsByFriends(String jwt);
-    Post updatePostContent(int postId, String content, MultipartFile file, String jwt) throws Exception;
+    Post updatePostContent(Long postId, String content, MultipartFile file, String jwt) throws Exception;
     List<Post> getPostsOrderedByReportCount();
 }
