+++ SocialMediaApp/SocialMediaApp/src/main/java/com/Licenta/SocialMediaApp/Service/ReportService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface ReportService {
     Report createReport(Report report);
-    void deleteReportById(int reportId, String jwt);
-    List<Report> findByPostId(int postId);
+    void deleteReportById(Long reportId, String jwt);
+    List<Report> findByPostId(Long postId);
 }
